# server/resources/openapi.json
{
"openapi":"3.0.3",
"info":{"title": "Focus Memo HTTP API", "version": "0.1.0", "description": "T1 minimal endpoints. Images are separate files. Full CRUD and desktop synchronization are not included yet."},
"servers":[{"url": "/"}],
"security":[{"bearerAuth": []}],
"paths":{"/api/v1/sync/changes":{"get":{"operationId":"getChanges","summary":"Fetch paged memo changes after a dataset-bound cursor","parameters":[{"in":"query","name":"cursor","required":true,"schema":{"type":"string","pattern":"^[0-9]+$"}},{"in":"query","name":"through","required":false,"description":"High-water mark from the first page; keep it unchanged while fetching the remaining pages","schema":{"type":"string","pattern":"^[0-9]+$"}},{"in":"query","name":"limit","schema":{"type":"integer","minimum":1,"maximum":10,"default":10}},{"in":"header","name":"X-Focus-Dataset-ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Changes page including tombstones and referenced ready attachment descriptors","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePage"}}}},"401":{"description":"Authentication required"},"409":{"description":"Dataset mismatch"},"422":{"description":"Invalid cursor, high-water mark or page size"},"503":{"description":"Unavailable; retry the same cursor"}}}},
"/healthz":{"get":{"operationId":"health","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string"}}}}}},"401":{"description":"Missing or invalid Bearer token"},"503":{"description":"Unavailable; retry writes with the same Idempotency-Key"}},"security":[]}},
"/api/v1/meta":{"get":{"operationId":"getMeta","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"datasetId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"integer"}}}}}},"401":{"description":"Missing or invalid Bearer token"},"503":{"description":"Unavailable; retry writes with the same Idempotency-Key"}}}},
"/api/v1/memos":{"post":{"operationId":"createMemo","responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receipt"}}}},"401":{"description":"Missing or invalid Bearer token"},"503":{"description":"Unavailable; retry writes with the same Idempotency-Key"},"409":{"description":"Idempotency mismatch or existing memo ID"},"413":{"description":"Body, image size or pixel limit exceeded"},"415":{"description":"Unsupported media type or invalid image"},"422":{"description":"Invalid fields, missing idempotency key or attachment"}},"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"header","name":"X-Focus-Dataset-ID","required":false,"description":"Desktop sync must send its fixed dataset binding; checked inside mutation transaction before any write. Omit only for unbound manual clients.","schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMemo"}}}}},"get":{"summary":"List/filter non-purged memos","parameters":[{"in":"query","name":"status","schema":{"type":"string","enum":["not_started","in_progress","completed"]}},{"in":"query","name":"priority","schema":{"type":"string","enum":["important","unimportant"]}},{"in":"query","name":"tag","schema":{"type":"string","maxLength":1024}},{"in":"query","name":"q","schema":{"type":"string","maxLength":4096}},{"in":"query","name":"deleted","schema":{"type":"boolean","default":false}},{"in":"query","name":"limit","schema":{"type":"integer","minimum":1,"maximum":500,"default":100}},{"in":"query","name":"offset","schema":{"type":"integer","minimum":0,"maximum":1000000,"default":0}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoList"}}}},"401":{"description":"Authentication required"},"409":{"description":"version_conflict, idempotency_mismatch, memo_purged or invalid_transition"},"422":{"description":"Invalid fields or baseVersion"},"503":{"description":"Unavailable; retry unchanged operation"}}}},
"/api/v1/memos/{id}":{"get":{"operationId":"getMemo","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Memo"}}}},"401":{"description":"Missing or invalid Bearer token"},"503":{"description":"Unavailable; retry writes with the same Idempotency-Key"},"404":{"description":"Not found"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]},"patch":{"summary":"Versioned memo update/restore","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"header","name":"X-Focus-Dataset-ID","required":false,"description":"Desktop sync must send its fixed dataset binding; checked inside mutation transaction before any write. Omit only for unbound manual clients.","schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchMemo"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receipt"}}}},"401":{"description":"Authentication required"},"409":{"description":"version_conflict, idempotency_mismatch, memo_purged or invalid_transition"},"422":{"description":"Invalid fields or baseVersion"},"503":{"description":"Unavailable; retry unchanged operation"}}},"delete":{"summary":"Versioned soft delete or purge; purge requires an already deleted memo","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"query","name":"baseVersion","required":true,"schema":{"type":"string","pattern":"^[1-9][0-9]*$","description":"Unsigned 64-bit integer transmitted as decimal string"}},{"in":"query","name":"permanent","schema":{"type":"boolean","default":false}},{"in":"header","name":"X-Focus-Dataset-ID","required":false,"description":"Desktop sync must send its fixed dataset binding; checked inside mutation transaction before any write. Omit only for unbound manual clients.","schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receipt"}}}},"401":{"description":"Authentication required"},"409":{"description":"version_conflict, idempotency_mismatch, memo_purged or invalid_transition"},"422":{"description":"Invalid fields or baseVersion"},"503":{"description":"Unavailable; retry unchanged operation"}}}},
"/api/v1/attachments":{"post":{"operationId":"uploadAttachment","responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Attachment"}}}},"401":{"description":"Missing or invalid Bearer token"},"503":{"description":"Unavailable; retry writes with the same Idempotency-Key"},"409":{"description":"Idempotency mismatch or existing memo ID"},"413":{"description":"Body, image size or pixel limit exceeded"},"415":{"description":"Unsupported media type or invalid image"},"422":{"description":"Invalid fields, missing idempotency key or attachment"}},"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","format":"uuid"}},{"in":"header","name":"X-Focus-Dataset-ID","required":false,"description":"Desktop sync must send its fixed dataset binding; checked inside mutation transaction before any write. Omit only for unbound manual clients.","schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["file"],"additionalProperties":false,"properties":{"file":{"type":"string","format":"binary"}}}}}}}},
"/api/v1/attachments/{id}":{"get":{"operationId":"getAttachment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Authenticated original image","content":{"image/png":{"schema":{"type":"string","format":"binary"}},"image/jpeg":{"schema":{"type":"string","format":"binary"}},"image/webp":{"schema":{"type":"string","format":"binary"}},"image/gif":{"schema":{"type":"string","format":"binary"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"503":{"description":"File unavailable"}}}},
"/api/v1/sync/snapshot":{"get":{"summary":"Consistent complete metadata snapshot including tombstones and ready attachment descriptions","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Snapshot"}}}},"401":{"description":"Authentication required"},"409":{"description":"version_conflict, idempotency_mismatch, memo_purged or invalid_transition"},"422":{"description":"Invalid fields or baseVersion"},"503":{"description":"Unavailable; retry unchanged operation"}}}}
},
"components":{
"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},
"schemas":{"CreateMemo":{"type":"object","additionalProperties":false,"properties":{"id":{"type":"string","format":"uuid"},"content":{"type":"string"},"type":{"type":"string","enum":["text","image"]},"priority":{"type":"string","enum":["important","unimportant"]},"status":{"type":"string","enum":["not_started","in_progress","completed"]},"tags":{"type":"array","items":{"type":"string"}},"attachmentIds":{"type":"array","items":{"type":"string","format":"uuid"}},"createdAt":{"type":"string","format":"date-time","description":"Optional historical display timestamp for approved first import; not used for conflict resolution"},"updatedAt":{"type":"string","format":"date-time","description":"Optional historical display timestamp for approved first import; not used for conflict resolution"},"completedAt":{"type":"string","format":"date-time","description":"Optional historical display timestamp for approved first import; not used for conflict resolution","nullable":true},"deviceId":{"type":"string","maxLength":128}}},"Receipt":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"appliedVersion":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"completedAt":{"type":"string","nullable":true},"deleted":{"type":"boolean"},"purged":{"type":"boolean"}}},"Memo":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"content":{"type":"string"},"type":{"type":"string","enum":["text","image"]},"priority":{"type":"string","enum":["important","unimportant"]},"status":{"type":"string","enum":["not_started","in_progress","completed"]},"tags":{"type":"array","items":{"type":"string"}},"attachmentIds":{"type":"array","items":{"type":"string","format":"uuid"}},"version":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"completedAt":{"type":"string","nullable":true},"deviceId":{"type":"string"},"deleted":{"type":"boolean"},"purged":{"type":"boolean"}}},"Attachment":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"mimeType":{"type":"string"},"size":{"type":"integer","format":"int64"},"sha256":{"type":"string"},"createdAt":{"type":"string"}}},"PatchMemo":{"type":"object","additionalProperties":false,"required":["baseVersion"],"minProperties":2,"properties":{"baseVersion":{"type":"string","pattern":"^[1-9][0-9]*$","description":"Unsigned 64-bit integer transmitted as decimal string"},"content":{"type":"string"},"type":{"type":"string","enum":["text","image"]},"priority":{"type":"string","enum":["important","unimportant"]},"status":{"type":"string","enum":["not_started","in_progress","completed"]},"tags":{"type":"array","items":{"type":"string"}},"attachmentIds":{"type":"array","items":{"type":"string","format":"uuid"}},"deleted":{"type":"boolean","enum":[false],"description":"Restore a soft-deleted memo"}}},"Meta":{"type":"object","required":["datasetId","schemaVersion"],"properties":{"datasetId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"integer","enum":[1]}}},"MemoList":{"type":"object","required":["memos"],"properties":{"memos":{"type":"array","items":{"$ref":"#/components/schemas/Memo"}}}},"MemoChange":{"type":"object","required":["sequence","memo"],"properties":{"sequence":{"type":"string","pattern":"^[0-9]+$"},"memo":{"$ref":"#/components/schemas/Memo"}}},"ChangePage":{"type":"object","required":["datasetId","schemaVersion","cursor","highWater","hasMore","changes","attachments"],"properties":{"datasetId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"integer","enum":[1]},"cursor":{"type":"string","pattern":"^[0-9]+$"},"highWater":{"type":"string","pattern":"^[0-9]+$"},"hasMore":{"type":"boolean"},"changes":{"type":"array","items":{"$ref":"#/components/schemas/MemoChange"}},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/Attachment"}}}},"Snapshot":{"type":"object","required":["datasetId","schemaVersion","changeCursor","complete","totalRows","memos","attachments"],"properties":{"datasetId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"integer","enum":[1]},"changeCursor":{"type":"string","pattern":"^[0-9]+$"},"complete":{"type":"boolean","enum":[true]},"totalRows":{"type":"integer","minimum":0},"memos":{"type":"array","items":{"$ref":"#/components/schemas/Memo"}},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/Attachment"}}}}}
}
}